This checkin includes code to make the IEMsgList work again for chat/room history.
In order for IE to be included in the compile, USE_TWEBBROWSER must be defined.  The current default is to NOT include the TWebBrowser in the build for chat/room History.txt

After TWebBrowser has been compiled into the EXE, then the msglist_type must be set (via default.xml resource or branding or prefs) to actually use IE as the chat/room history.  Otherwise, even though TWebBrowser is compiled in the executable, it will still use RTF.

Special note: TWebBrowser uses far more memory then using RTF.